backend completed

+++ oerms-backend/attempt-service/src/main/java/com/oerms/attempt/dto/AttemptAnswerDTO.java
@@ -19,7 +19,7 @@ public class AttemptAnswerDTO {
     private String answerText;
 
     private Integer marksAllocated;
-    private Integer timeSpentSeconds;
+    private Long timeSpentSeconds;
 
     private Boolean flagged;
     private LocalDateTime answeredAt;
